+++ libstreetmap/src/m1_more.h
@@ -49,7 +49,7 @@ extern std::unordered_map<OSMID, const OSMWay*> WayID_Way;
 extern std::unordered_map<OSMID, double> WayID_length;
 
 // the maximum speed_limit in the city
-extern float max_speed_limit ;
+extern double max_speed_limit ;
 
 int find_closest_POI(LatLon my_position);
